// src/app/models/company.ts
export class Company {
  _id: number;
  isWinner: any;
  tenderNumber: string;
  companyName: string;
  companyRegistrationNo: string;
  companyPhoneNumber: string;
  companyAddress: string;
  directors: string;
  winningcount: number;
  awardedPoint: number;
  createDate: string;
  tenderRef: string;
  companyEmail: string;
}